make some pre-jdk-8 tests more direct
