+++ test/files/pos/sammy_java8/F.java
@@ -1,0 +1,6 @@
+public interface F<T, U> {
+    U apply(T t);
+    default void yadayada() {
+        throw new UnsupportedOperationException("yadayada");
+    }
+}
