+++ test/files/pos/sammy_java8/Test.scala
@@ -1,0 +1,4 @@
+class T {
+  def app[T, U](x: T)(f: F[T, U]): U = f(x)
+  app(1)(x => List(x))
+}
