+++ test/files/pos/t7398/Test.scala
@@ -1,0 +1,5 @@
+class Test extends Iterator[String] {
+  def hasNext = true
+  def next() = ""
+  def test = this.remove()
+}
