+++ test/files/pos/t8852/Interface.java
@@ -1,0 +1,5 @@
+public interface Interface {
+  public static int staticMethod() {
+    return 42;
+  }
+}
