+++ test/files/pos/t8852/Test.scala
@@ -1,0 +1,5 @@
+object Test {
+  val x: Int = Interface.staticMethod()
+}
+
+class C extends Interface // expect no errors about unimplemented members.
